# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: React Azure Active Directory API
  version: 1.0.0
  description: An Azure Active Directory setup using a React frontend and a Node Express backend
  contact:
    name: Steve Renner
    url: https://github.com/fsrenner
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        aadId:
          type: string
          description: The Azure Active Directory Id
          example: "fdc1d8b4-e667-4f9e-853a-0b2fdf742ac6"
        createdAt:
          type: string
          description: The timestamp when the entity was created
          example: "2020-11-30T16:52:27.011Z"
        updatedAt:
          type: string
          description: The timestamp when the entity was created
          example: "2020-11-30T16:53:07.409Z"
        emailAddress:
          type: string
          description: The user email address
          example: "f.steve.renner@gmail.com"
        firstName:
          type: string
          description: The user first name
          example: "Frederick"
        lastName:
          type: string
          description: The user last name
          example: "Renner"
        id:
          type: number
          description: The entity id
          example: 1002
...
    NotLoggedIn:
      type: object
      properties:
        message:
          type: string
          description: User not logged in and not passing bearer token
          example: "Please login in order to access the application"
...
paths:
...
  /v1/users:
    get:
      security:
      - bearerAuth: []  
      summary: Fetch all users
      operationId: getUsers
      tags:
        - Users
      responses:
        '200':
          description: array of user objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotLoggedIn'
    post:
      security:
      - bearerAuth: []  
      summary: Create new user
      operationId: createUser
      requestBody:
        description: parameters necessary to create a new user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emailAddress:
                  type: string
                  description: The user email address
                  example: "f.steve.renner@gmail.com"
                firstName:
                  type: string
                  description: The user first name
                  example: "Frederick"
                lastName:
                  type: string
                  description: The user last name
                  example: "Renner"
      tags:
        - Users
      responses:
        '201':
          description: The created user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotLoggedIn'
...
  /v1/users/id:
    get:
      security:
      - bearerAuth: []  
      summary: Fetch all user ids
      operationId: getUserIds
      tags:
        - Users
      responses:
        '200':
          description: array of user ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  example: 1
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotLoggedIn'